requirements.txt
flask
flask_restful
flask-wtf
sqlalchemy
flask-sqlalchemy
sqlalchemy-migrate
